refactor msvc module infrastructure
